Updated the UI and simplified flow service
